refactor(component): refactor FindTemplatePage for code readability
