+++ apps/web/templates/form/index.ts
@@ -1,2 +1,3 @@
 export * from './ConfigForm';
 export * from './LoginForm';
+export * from './FindTemplateSearchForm';
